Extract tags toggle code to separate file
